wip: create useIsHydrated hook
