+++ packages/headless/src/components/Portal/index.tsx
@@ -1,25 +1,16 @@
-import { type FC, type PropsWithChildren, useSyncExternalStore } from 'react'
+import { type FC, type PropsWithChildren } from 'react'
 import { createPortal } from 'react-dom'
+import { useIsHydrated } from '../../hooks/useIsHydrated/index.js'
 
 type PortalProps = PropsWithChildren<{
     target?: HTMLElement
 }>
 
-const noopSubscribe = () => {
-    return () => {}
-}
-
 export const Portal: FC<PortalProps> = ({
     children,
     target = typeof window !== 'undefined' ? document?.body : null,
 }) => {
-    // ensures we are client side rendering before using createPortal
-    // which is not supported on the server
-    const hydrated = useSyncExternalStore(
-        noopSubscribe,
-        () => true, // snapshot always returns true because at this point we're on the client
-        () => false, // server snapshot alwasy returns false because at this point we're on the server
-    )
-
+    const hydrated = useIsHydrated()
+    
     return hydrated && target ? createPortal(children, target) : null
 }
